+++ tests/serializers/test_hfml_serializer.py
@@ -1,10 +1,8 @@
-import tempfile
 from pathlib import Path
 
 import pytest
-from yaml import serialize
 
-from openpecha.serializers import EditorSerializer, EpubSerializer, HFMLSerializer
+from openpecha.serializers import HFMLSerializer
 
 
 # For testing read_base_layer, add_chars and apply_layer
@@ -27,21 +25,3 @@ def test_hfml_serializer():
     expected_hfml_vol2 = Path('./tests/data/serialize/hfml/expected_hfml/P000003/v002.txt').read_text(encoding='utf-8')
     assert hfml_results['v001'] == expected_hfml_vol1
     assert hfml_results['v002'] == expected_hfml_vol2
-
-
-def test_hfml_2_tsadra_serializer(opf_path):
-    with tempfile.TemporaryDirectory() as tmpdirname:
-        serializer = EpubSerializer(opf_path)
-        out_fn = serializer.serialize(output_path=tmpdirname)
-        assert Path(out_fn).name == "P000100.epub"
-
-
-def test_editor_serializer():
-    opf_path = "./tests/data/serialize/tsadra/P000801/P000801.opf"
-    expected_output = Path(
-        "./tests/data/serialize/tsadra/editor_serializer_output.html"
-    ).read_text(encoding="utf-8")
-    serializer = EditorSerializer(opf_path)
-    for base_name, result in serializer.serialize():
-        assert base_name
-        assert result == expected_output